refactor: switch-theme
