+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './helpers';
+export * from './themes';
 export * from './utils';
